tests: move fixtures to global conftest
